Extract some modules and reorder code for readability
